revise to match only return value
I added number of exception to count the occurrence of '\0' at %c and to
add it to return value. This way fixed return value but still test
produce difference.

I tried cat -e option to see if the stdout is the same. I found ^@ at
the poistion of '\0' so I need to distinguish string of 00 and 0,
impossible.

I think I need to change structure of list so it can save strlen of 00
so to print first 0 and return correct character count.